# Allow creating Status with no error message