Summary:
Status::UserError("") internally calls fmt::format("") and this call shows up pretty high on the profile (16%).

 {F1647622447}

Allow creating Status objects without passing any error details, e.g. Status::UserError().

This change helps speed up TRY(CAST(...)) benchmark by ~20%.

Before:

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast##try_cast_invalid_empty_input                          2.27ms    439.78
cast##tryexpr_cast_invalid_empty_input                      5.93ms    168.61
cast##try_cast_invalid_nan                                  5.64ms    177.35
cast##tryexpr_cast_invalid_nan                              9.44ms    105.94
```

After:

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast##try_cast_invalid_empty_input                          2.16ms    463.62
cast##tryexpr_cast_invalid_empty_input                      4.29ms    232.95
cast##try_cast_invalid_nan                                  5.47ms    182.83
cast##tryexpr_cast_invalid_nan                              7.76ms    128.81
```

Differential Revision: D57728340